+++ server/utils/sendinblue/index.js
@@ -1,12 +1,13 @@
 import https from 'https';
 import config from '../../config/environment';
 
-export const sendTransactionalEmail = message => new Promise((resolve, reject) => {
-  const data = JSON.stringify(message);
+// POST or PUT submission to SendInBlue
+const sibSubmit = (method, path, content) => new Promise((resolve, reject) => {
+  const data = JSON.stringify(content);
   const options = {
     hostname: 'api.sendinblue.com',
-    path: '/v3/smtp/email',
-    method: 'POST',
+    path,
+    method,
     headers: {
       'api-key': config.mail.apiKey,
       'Content-Type': 'application/json',
@@ -16,7 +17,7 @@ export const sendTransactionalEmail = message => new Promise((resolve, reject) =
   const req = https.request(options, res => {
     res.on('data', buffer => {
       if(res.statusCode === 201) return resolve(buffer.toString('utf8'));
-      reject('Unexpected response from SendInBlue');
+      reject(`Unexpected response (${res.statusCode}) from SendInBlue`);
     });
     req.on('error', err => {
       console.error('Error communicating with SendInBlue', err);
@@ -26,3 +27,31 @@ export const sendTransactionalEmail = message => new Promise((resolve, reject) =
   req.write(data);
   req.end();
 });
+
+/*
+  Sample contactInfo format
+  {
+    updateEnabled: true,
+    email: 'foo@example.com',
+    emailBlacklisted: false
+    attributes: {
+      NAME: 'First',
+      SURNAME: 'Last',
+    }
+  }
+*/
+export const sibContactUpsert = contactInfo => sibSubmit('POST', '/v3/contacts', contactInfo);
+
+export const sibOptOut = email => sibSubmit('PUT', `/v3/contacts/${email}`, { emailBlacklisted: true });
+
+/*
+  Sample message format
+  {
+    sender: [{ email: 'foo@example.com', name: 'Something' }], //sender: config.mail.sender,
+    to: [{ email: 'foo@example.com', name: 'Something' }],
+    subject: 'My Subject',
+    htmlContent: 'This is test content',
+    tags: ['myTag1', 'myTag2']
+  }
+*/
+export const sibSendTransactionalEmail = message => sibSubmit('POST', '/v3/smtp/email', message);